Compute profits in both directions
Previously, we only considered the case where the trader made a profit
in the 'natural' forward direction:

- open "swap x tok1 for n tok2",
- (target swap tok1 for tok2)
- close "swap n tok2 for y tok1"

where y-x is the profit. This makes most sense when tok1 is
ETH.

But it is also possible to take a profit in the other direction:

- open "swap n tok1 for x tok2",
- (target swap tok1 for tok2)
- close "swap y tok2 for n tok1"

where x-y is the profit. This makes most sense when tok2 is ETH, and
the transaction being sandwiched is swapExactTokensForEth.