Rearrange profits
Previously, '.profit' and '.profit2' held the forward and
backward profit. The backward profit was dropped if zero but the
forward wasn't, which is inconsistent. Furthermore, the distincgion of
"forward" and "backward" isn't really important for the FE app at this
point.

Consequently, this PR changes things to drop any zero
profit (regardless of forward or backward). As a result, the backward
profit can be in '.profit'.