Handle large wallets
Some wallets have large amounts of transactions that are too large to
hande with our current in-line approach. ("too large" means for example
an exchange wallet we tried out that had 2m transactions).

As a stopgap, this commit adds a limit to the number of logs we
fetch. It also adds some handling for a case where getPastLogs is
returning null (which it shouldn't, but ah well).